Do not reuse propagators
Propagator reuse is tricky with parameterized pulses. For example, when
using PWC propagators such as Cheby, the propagator determines the
amplitudes of the control at the midpoins of the time grid as internal
parameters during initialization. These are not connected to the
analytic parameters in the control, and updating the analytic parameters
would not be reflected in a subsquent propagation.

The only way to get around this in general is to do every propagation
witha fresh propagator, even though that implies some overhead.